fixed code for using BigDecimal instead of Double

+++ src/main/java/com/example/demo/init/Init.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,12 +21,12 @@ public class Init {
     @PostConstruct
     private void postConstruct(){
         List<Good> goodList = new ArrayList<>();
-        goodList.add(new Good("Яблоки краснодарские",98.7));
-        goodList.add(new Good("Яблоки Ред  Чиф",99.5));
-        goodList.add(new Good("Мандарины",124.5));
-        goodList.add(new Good("Груши Форелл",124.5));
-        goodList.add(new Good("Груши Пакхам",99.5));
-        goodList.add(new Good("Coca-Cola  0,5 л",95.0));
+        goodList.add(new Good("Яблоки краснодарские", BigDecimal.valueOf(98.70)));
+        goodList.add(new Good("Яблоки Ред  Чиф",BigDecimal.valueOf(99.5)));
+        goodList.add(new Good("Мандарины",BigDecimal.valueOf(124.5)));
+        goodList.add(new Good("Груши Форелл",BigDecimal.valueOf(124.5)));
+        goodList.add(new Good("Груши Пакхам",BigDecimal.valueOf(99.5)));
+        goodList.add(new Good("Coca-Cola  0,5 л",BigDecimal.valueOf(95.0)));
         goodList.forEach(good -> goodRepository.saveAndFlush(good));
